requirements.txt
Flask==2.0.1
python-dotenv==0.19.0
google-generativeai==0.1.0
matplotlib==3.4.3